fixed I18n, move MaintenancePage to correct package
